Update File.cs Bug Fix
Bug 1:
If the file extension consists of uppercase letters, when the allowed file types are checked, the error causes to  throw an error.
To fix this, the file extension has been changed to lowercase. Now, the file will save with this extension.

Bug 2:
When the validation conditions are not met, the delete method that run inside brackets, had taken wrong parameter. The parameter must be relative path but it's not. Because of that it has changed with right parameter which that is the variable named link.

P.S.: I didn't check for netcore but probably it will work well